chore(db): rename user to users
